Add AppBar component
